assist dto

+++ src/main/java/br/com/fiap/abctechservice/model/OrderLocation.java
@@ -6,7 +6,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 import javax.persistence.*;
-import java.sql.Date;
+import java.util.Date;
 
 
 @Entity
@@ -20,8 +20,7 @@ public class OrderLocation {
     @Column(name = "id", nullable = false)
     private Long id;
 
-    private Integer latitude;
-    private Integer longitude;
+    private Double latitude;
+    private Double longitude;
     private Date date;
 }
-
